+++ cmd/immuclient/commands/print.go
@@ -1,0 +1,120 @@
+package commands
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/codenotary/immudb/pkg/api"
+	"github.com/codenotary/immudb/pkg/api/schema"
+	"github.com/codenotary/immudb/pkg/client"
+	"github.com/codenotary/immudb/pkg/store"
+)
+
+func printItem(key []byte, value []byte, message interface{}) {
+	var index, ts uint64
+	var verified, isVerified bool
+	var hash []byte
+	switch m := message.(type) {
+	case *schema.Index:
+		index = m.Index
+		dig := api.Digest(index, key, value)
+		hash = dig[:]
+	case *client.VerifiedIndex:
+		index = m.Index
+		dig := api.Digest(index, key, value)
+		hash = dig[:]
+		verified = m.Verified
+		isVerified = true
+	case *schema.Item:
+		key = m.Key
+		value = m.Value
+		index = m.Index
+		hash = m.Hash()
+	case *schema.StructuredItem:
+		key = m.Key
+		value = m.Value.Payload
+		ts = m.Value.Timestamp
+		index = m.Index
+		hash, _ = m.Hash()
+	case *client.VerifiedItem:
+		key = m.Key
+		value = m.Value
+		index = m.Index
+		ts = m.Time
+		verified = m.Verified
+		isVerified = true
+		me, _ := schema.Merge(value, ts)
+		dig := api.Digest(index, key, me)
+		hash = dig[:]
+
+	}
+	if !isVerified {
+		fmt.Printf("index:		%d\nkey:		%s\nvalue:		%s\nhash:		%x \ntime:		%s \n",
+			index,
+			key,
+			value,
+			hash,
+			time.Unix(int64(ts), 0))
+		return
+	}
+	fmt.Printf("index:		%d\nkey:		%s\nvalue:		%s\nhash:		%x \ntime:		%s \nverified:	%t \n",
+		index,
+		key,
+		value,
+		hash,
+		time.Unix(int64(ts), 0),
+		verified)
+}
+
+func printSetItem(set []byte, rkey []byte, score float64, message interface{}) {
+	var index uint64
+	var verified, isVerified bool
+	switch m := message.(type) {
+	case *schema.Index:
+		index = m.Index
+	case *client.VerifiedIndex:
+		index = m.Index
+		verified = m.Verified
+		isVerified = true
+	}
+	key, err := store.SetKey(rkey, set, score)
+	if err != nil {
+		fmt.Print(err.Error())
+	}
+	if !isVerified {
+		fmt.Printf("index:		%d\nset:		%s \nkey:		%s \nscore:		%f \nvalue:		%s \nhash:		%x \n",
+			index,
+			set,
+			key,
+			score,
+			rkey,
+			api.Digest(index, key, rkey))
+		return
+	}
+	fmt.Printf("index:		%d\nset:		%s\nkey:		%s\nscore:		%f\nvalue:		%s\nhash:		%x\nverified:	%t\n",
+		index,
+		set,
+		key,
+		score,
+		rkey,
+		api.Digest(index, key, rkey),
+		verified)
+}
+
+func printRoot(root *schema.Root) {
+	if root.Root == nil {
+		fmt.Printf("Immudb is empty\n")
+		return
+	}
+	fmt.Printf("index:		%d\nhash:		%x\n", root.Index, root.Root)
+}
+
+func printByIndex(item *schema.StructuredItem) {
+	dig, _ := item.Hash()
+	fmt.Printf("index:		%d\nkey:		%s\nvalue:		%s\nhash:		%x\ntime:		%s\n",
+		item.Index,
+		item.Key,
+		item.Value,
+		dig,
+		time.Unix(int64(item.Value.Timestamp), 0))
+}
