+++ pkg/client/types.go
@@ -1,12 +1,9 @@
 /*
 Copyright 2019-2020 vChain, Inc.
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
 	http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -83,8 +80,10 @@ type VerifiedIndex struct {
 func (vi *VerifiedIndex) Reset() { *vi = VerifiedIndex{} }
 
 func (vi *VerifiedIndex) String() string { return proto.CompactTextString(vi) }
+
 // ProtoMessage ...
 func (*VerifiedIndex) ProtoMessage() {}
+
 // Reset ...
 func (vi *VerifiedItem) Reset() { *vi = VerifiedItem{} }
